# openapi/pileofmedia-openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "pileofmedia",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.3.0"
  },
  "servers": [
    {
      "url": "https://pileofmedia.com/api/",
      "description": "Production server"
    },
    {
      "url": "https://Localhost:5000/api/",
      "description": "Localhost"
    }
  ],
  "paths": {
...
    "/books": {
      "get": {
        "tags": [
          "Books"
        ],
        "operationId": "get_items",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "The language of the items",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "The current page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "The amount of items to query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "maximum": 50,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned books",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooksTotal"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Books"
        ],
        "operationId": "create_book",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateBook"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Book successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/characters": {
      "get": {
        "tags": [
          "Characters"
        ],
        "operationId": "get_items",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "The language of the items",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "The current page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "The amount of items to query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "maximum": 50,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned characters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharactersTotal"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Characters"
        ],
        "operationId": "create_item",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateCharacter"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Character successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Character"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/franchises": {
      "get": {
        "tags": [
          "Franchises"
        ],
        "operationId": "get_items",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "The language of the items",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "The current page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "The amount of items to query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "maximum": 50,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned franchises",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FranchisesTotal"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Franchises"
        ],
        "operationId": "create_item",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFranchise"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Franchise successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Franchise"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/genres": {
      "get": {
        "tags": [
          "Genres"
        ],
        "operationId": "get_items",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "The language of the items",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "The current page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "The amount of items to query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "maximum": 50,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned genres",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenresTotal"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Genres"
        ],
        "operationId": "create_item",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGenre"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Genre successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Genre"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/people": {
      "get": {
        "tags": [
          "People"
        ],
        "operationId": "get_items",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "The language of the items",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "The current page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "The amount of items to query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "maximum": 50,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned people",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeopleTotal"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "People"
        ],
        "operationId": "create_item",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreatePerson"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Person successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/roles": {
      "get": {
        "tags": [
          "Roles"
        ],
        "operationId": "get_items",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "The language of the items",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "The current page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "The amount of items to query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "maximum": 50,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned roles",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RolesTotal"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Roles"
        ],
        "operationId": "create_item",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRole"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Role successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/themes": {
      "get": {
        "tags": [
          "Themes"
        ],
        "operationId": "get_items",
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "The language of the items",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "The current page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "The amount of items to query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "maximum": 50,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned themes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThemesTotal"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Themes"
        ],
        "operationId": "create_item",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTheme"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Theme successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Theme"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "get_items",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "The current page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "The amount of items to query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true,
              "maximum": 50,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersTotal"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Book": {
        "type": "object",
        "required": [
          "id",
          "slug",
          "genres",
          "themes",
          "involved",
          "characters",
          "translations",
          "statistic"
        ],
        "properties": {
          "characters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookCharacter"
            }
          },
          "franchise": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Franchise"
              }
            ],
            "nullable": true
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Genre"
            }
          },
          "id": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "involved": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Involved"
            }
          },
          "published": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "slug": {
            "$ref": "#/components/schemas/Slug"
          },
          "statistic": {
            "$ref": "#/components/schemas/BookStatistic"
          },
          "themes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Theme"
            }
          },
          "translations": {
            "$ref": "#/components/schemas/BookAvailableTranslations"
          }
        }
      },
...
      "BooksTotal": {
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Book"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "Character": {
        "type": "object",
        "required": [
          "id",
          "translations"
        ],
        "properties": {
          "birthday": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "height_cm": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "minimum": 0
          },
          "id": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "image": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Image"
              }
            ],
            "nullable": true
          },
          "translations": {
            "$ref": "#/components/schemas/CharacterAvailableTranslations"
          }
        }
      },
...
      "CharactersTotal": {
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Character"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "CreateBook": {
        "type": "object",
        "required": [
          "book",
          "covers"
        ],
        "properties": {
          "book": {
            "$ref": "#/components/schemas/CreateBookData"
          },
          "covers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateImage"
            }
          }
        }
      },
...
      "CreateCharacter": {
        "type": "object",
        "required": [
          "character"
        ],
        "properties": {
          "character": {
            "$ref": "#/components/schemas/CreateCharacterData"
          },
          "image": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CreateImage"
              }
            ],
            "nullable": true
          }
        }
      },
...
      "CreateFranchise": {
        "type": "object",
        "required": [
          "translations"
        ],
        "properties": {
          "translations": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/CreateFranchiseTranslation"
            }
          }
        }
      },
...
      "CreateGenre": {
        "type": "object",
        "required": [
          "translations"
        ],
        "properties": {
          "translations": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/CreateGenreTranslation"
            }
          }
        }
      },
...
      "CreateImage": {
        "type": "string",
        "format": "binary"
      },
      "CreatePerson": {
        "type": "object",
        "required": [
          "person"
        ],
        "properties": {
          "image": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CreateImage"
              }
            ],
            "nullable": true
          },
          "person": {
            "$ref": "#/components/schemas/CreatePersonData"
          }
        }
      },
...
      "Franchise": {
        "type": "object",
        "required": [
          "id",
          "translations"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "translations": {
            "$ref": "#/components/schemas/FranchiseAvailableTranslations"
          }
        }
      },
...
      "FranchisesTotal": {
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Franchise"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "Genre": {
        "type": "object",
        "required": [
          "id",
          "translations"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "translations": {
            "$ref": "#/components/schemas/GenreAvailableTranslations"
          }
        }
      },
...
      "GenresTotal": {
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Genre"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "PeopleTotal": {
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Person"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "Person": {
        "type": "object",
        "required": [
          "id",
          "name",
          "translations"
        ],
        "properties": {
          "birthday": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "first_name": {
            "type": "string",
            "nullable": true
          },
          "height_cm": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "minimum": 0
          },
          "id": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "image": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Image"
              }
            ],
            "nullable": true
          },
          "last_name": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "translations": {
            "$ref": "#/components/schemas/PersonAvailableTranslations"
          }
        }
      },
...
      "RolesTotal": {
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "Slug": {
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "user_token": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}